BindingGroup rename in samples/tests
