+++ src/DotVVM.Samples.Common/ViewModels/FeatureSamples/BindableCssStyles/BindableCssStylesViewModel.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using DotVVM.Framework.ViewModel;
 
-namespace DotVVM.Samples.BasicSamples.ViewModels.FeatureSamples.BindableStyleProperties
+namespace DotVVM.Samples.BasicSamples.ViewModels.FeatureSamples.BindableCssStyles
 {
-    public class BindableStylePropertiesViewModel : DotvvmViewModelBase
+    public class BindableCssStylesViewModel : DotvvmViewModelBase
     {
         public int Width { get; set; } = 50;
         public int FontSize { get; set; } = 14;
